ext4/048: Fix hangup due to no free inodes
We currently mkfs a 128MB filesystem, which gives use ~2048 free inodes
on 64k blocksize. The test then keeps adding new files to a directory to
trigger an htree split. For 64k this takes more than the total free
inodes, which causes touch to return -ENOSPC. This leads to the while
loop in induce_node_split() to never finish.

To fix this:
1. Format a 1G FS which gives us atleast 16K inodes to work with.
2. _fail if there's any error while trying to induce node split, so we
   dont get stuck in loop

Fixes: 466ddbfd1151 ("ext4: add test for ext4_dir_entry2 wipe")
Signed-off-by: Ojaswin Mujoo <ojaswin@linux.ibm.com>
Reviewed-by: Baokun Li <libaokun1@huawei.com>
Signed-off-by: Zorro Lang <zlang@kernel.org>