generic: test a scenario of power failure after renames and fsyncs
Test moving a directory to another location, create a file in the old
location of the directory and with the same name, fsync the file, then
move the file elsewhere and fsync again the file. After a power failure
we expect to be able to mount the fs and have the same content as before
the power failure.

This exercises a bug fixed by the following kernel patch for btrfs:

 "btrfs: don't log conflicting inode if it's a dir moved in the current transaction"

Signed-off-by: Filipe Manana <fdmanana@suse.com>
Reviewed-by: Darrick J. Wong <djwong@kernel.org>
Signed-off-by: Zorro Lang <zlang@kernel.org>