fsstress: allow multiple suboptions to -f
I got bitten by fsstress's argument parsing recently because it turns
out that if you do:

# fsstress -z -f creat=2,unlink=1

It will ignore everything after the '2' and worse yet it won't tell you
that it's done so unless you happen to pass -S to make it spit out the
frequency table.

Adapt process_freq to tokenize the argument string so that it can handle
a comma-separated list of key-value arguments.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>
Reviewed-by: Zorro Lang <zlang@redhat.com>
Signed-off-by: Zorro Lang <zlang@kernel.org>